arm64jit: Reuse code in I2R funcs.

+++ Common/Arm64Emitter.h
@@ -714,13 +714,13 @@ public:
 	void SUBI2R(ARM64Reg Rd, ARM64Reg Rn, u64 imm, ARM64Reg scratch = INVALID_REG);
 	void SUBSI2R(ARM64Reg Rd, ARM64Reg Rn, u64 imm, ARM64Reg scratch = INVALID_REG);
 
-	bool TryADDI2R(ARM64Reg Rd, ARM64Reg Rn, u32 imm);
-	bool TrySUBI2R(ARM64Reg Rd, ARM64Reg Rn, u32 imm);
-	bool TryCMPI2R(ARM64Reg Rn, u32 imm);
+	bool TryADDI2R(ARM64Reg Rd, ARM64Reg Rn, u64 imm);
+	bool TrySUBI2R(ARM64Reg Rd, ARM64Reg Rn, u64 imm);
+	bool TryCMPI2R(ARM64Reg Rn, u64 imm);
 
-	bool TryANDI2R(ARM64Reg Rd, ARM64Reg Rn, u32 imm);
-	bool TryORRI2R(ARM64Reg Rd, ARM64Reg Rn, u32 imm);
-	bool TryEORI2R(ARM64Reg Rd, ARM64Reg Rn, u32 imm);
+	bool TryANDI2R(ARM64Reg Rd, ARM64Reg Rn, u64 imm);
+	bool TryORRI2R(ARM64Reg Rd, ARM64Reg Rn, u64 imm);
+	bool TryEORI2R(ARM64Reg Rd, ARM64Reg Rn, u64 imm);
 
 	// Pseudo-instruction for convenience. PUSH pushes 16 bytes even though we only push a single register.
 	// This is so the stack pointer is always 16-byte aligned, which is checked by hardware!
